+++ changelog.md
@@ -1,3 +1,10 @@
+# 0.3.0.0
+
+* Fix bug in instance of `ToHttpApiData` for `Location`
+
+* Move functions and types of anticipated common use to module
+  `Web.Google.Maps.Common`
+
 # 0.2.0.0
 
 * Implementation of `signature` and custom marker icons
@@ -10,5 +17,5 @@
 # 0.1.0.0
 
 * Launch implementation. Not yet implemented: certain optional parameters
-  (language, region and signature); address locations; non-PNG image formats;
-  custom marker icons; and encoded polyline paths.
+  (`language`, `region` and `signature`); address locations; non-PNG image
+  formats; custom marker icons; and encoded polyline paths